pyproject.toml
[project]
name = "order-intake-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115",
    "google-adk>=1.31.0",
    "google-api-python-client>=2.140",
    "google-auth>=2.35",
    "google-auth-httplib2>=0.2",
    "google-auth-oauthlib>=1.2",
    "google-cloud-firestore>=2.27.0",
    "google-cloud-pubsub>=2.23",
    "llama-cloud>=2.1",
    "openpyxl>=3.1.5",
    "pydantic>=2",
    "pymupdf>=1.27.2.2",
    "python-dotenv>=1.2.2",
    "rapidfuzz>=3.9",
    "reportlab>=4.4.10",
    "rich>=15.0.0",
    "structlog>=25.5.0",
    "uvicorn[standard]>=0.32",
]

[dependency-groups]
dev = [
    "pytest>=8",
    "pytest-mock>=3.12",
    "pytest-asyncio>=0.23",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: tests that hit the live LlamaExtract API (requires LLAMA_CLOUD_API_KEY)",
    "firestore_emulator: tests that require a running Firestore emulator (FIRESTORE_EMULATOR_HOST)",
    "gmail_live: gated live Gmail integration (set GMAIL_LIVE_TEST=1)",
    "pubsub_emulator: requires the Cloud Pub/Sub emulator running locally (set PUBSUB_EMULATOR_HOST=localhost:8085)",
]
